requirements.txt
discord.py==1.7.3
Pillow==8.3.2
python-dotenv==0.19.0
web3==5.24.0